more conditional handling for AWS key env vars

+++ support/build/_util/deploy.sh
@@ -10,6 +10,11 @@ if [[ $# -lt 2 ]]; then
 	exit 2
 fi
 
+if [[ -z "${AWS_ACCESS_KEY_ID:-}" || -z "${AWS_SECRET_ACCESS_KEY:-}" ]]; then
+	echo '$AWS_ACCESS_KEY_ID or $AWS_SECRET_ACCESS_KEY not set!' >&2
+	exit 2
+fi
+
 # a helper (print_or_export_manifest_cmd) called in the script invoked by Bob will write to this if set
 export MANIFEST_CMD=$(mktemp -t "manifest.XXXXX")
 trap 'rm -rf $MANIFEST_CMD;' EXIT
